+++ libstuff/SRandom.h
@@ -1,0 +1,11 @@
+#pragma once
+
+// Random number generator class.
+class SRandom {
+private:
+    static mt19937_64* generator;
+    static uniform_int_distribution<uint64_t>* distribution64;
+    static void init();
+public:
+    static uint64_t rand64();
+};
