+++ libstuff/libstuff.h
@@ -210,19 +210,6 @@ struct SStopwatch {
     }
 };
 
-// Random number generator class.
-class SRandom {
-private:
-    static mt19937_64* generator;
-    static uniform_int_distribution<uint64_t>* distribution64;
-    static void init();
-public:
-    static uint64_t rand64();
-};
-
-// Compatibility function for the old random number generator (just `rand()`). Uses SRandom.
-uint64_t SRand64();
-
 // --------------------------------------------------------------------------
 // Signal stuff
 // --------------------------------------------------------------------------
@@ -901,3 +888,6 @@ struct STestTimer {
 #include "STCPNode.h"
 #include "SDataClient.h"
 #include "SHTTPSManager.h"
+
+// Other libstuff headers.
+#include "SRandom.h"
