+++ test/tests/LibStuffTest.cpp
@@ -283,13 +283,13 @@ struct LibStuff : tpunit::TestFixture
         string data = "";
         ASSERT_TRUE(SGZip(data).length() > 1);
 
-        data += (char)(SRand64() % 256);
+        data += (char)(SRandom::rand64() % 256);
         ASSERT_TRUE(SGZip(data).length() > 1);
 
         data = "";
         for (int i = 0; i < 10000000; i++)
         {
-            data += (char)(SRand64() % 256);
+            data += (char)(SRandom::randd64() % 256);
         }
         ASSERT_TRUE(SGZip(data).length() > 1);
